Move tests to dedicated directory
